requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
pyyaml>=6.0

# Recommendation systems libraries
implicit>=0.7.0
lightfm>=1.17
surprise>=1.1.3
lenskit>=0.14.0

# Deep learning (optional)
torch>=2.0.0
tensorflow>=2.13.0
tensorflow-recommenders>=0.7.0

# Geographic and distance calculations
geopy>=2.3.0
folium>=0.14.0

# Visualization and UI
matplotlib>=3.7.0
plotly>=5.15.0
streamlit>=1.25.0
seaborn>=0.12.0

# Development tools
tqdm>=4.65.0
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
pre-commit>=3.3.0

# Optional: Experiment tracking
mlflow>=2.5.0
wandb>=0.15.0

# Optional: Advanced features
faiss-cpu>=1.7.4
hnswlib>=0.7.0
sentence-transformers>=2.2.0
